+++ src/domain/usecases/ExportCorruptedEventReports.ts
@@ -1,10 +1,9 @@
-import { Async } from "domain/entities/Async";
 import { EventReport } from "domain/entities/EventReport";
 import { EventReportsExportRepository } from "domain/repositories/EventReportsExportRepository";
 
 export class ExportCorruptedEventReports {
     constructor(private eventReportsExportRepository: EventReportsExportRepository) {}
-    execute(eventReports: EventReport[]): Async<void> {
-        return this.eventReportsExportRepository.save(eventReports);
+    execute(eventReports: EventReport[], outputFile: string): void {
+        return this.eventReportsExportRepository.save(eventReports, outputFile);
     }
 }
